fix: refactoring csv views
主に CSV 関連のビューについてリファクタリング
